Move game palette-related stuff into Palette.{cpp.h}
