+++ src/openrct2/libopenrct2.vcxproj
@@ -266,6 +266,7 @@
     <ClInclude Include="drawing\LightFX.h" />
     <ClInclude Include="drawing\NewDrawing.h" />
     <ClInclude Include="drawing\PaletteIndex.h" />
+    <ClInclude Include="drawing\Palette.h" />
     <ClInclude Include="drawing\PaletteMap.h" />
     <ClInclude Include="drawing\PaletteType.h" />
     <ClInclude Include="drawing\PickupPeep.h" />
@@ -911,6 +912,7 @@
     <ClCompile Include="drawing\LightFX.cpp" />
     <ClCompile Include="drawing\Line.cpp" />
     <ClCompile Include="drawing\NewDrawing.cpp" />
+    <ClCompile Include="drawing\Palette.cpp" />
     <ClCompile Include="drawing\PaletteMap.cpp" />
     <ClCompile Include="drawing\PickupPeep.cpp" />
     <ClCompile Include="drawing\Rectangle.cpp" />
