+++ src/openrct2/scenario/Scenario.cpp
@@ -20,7 +20,7 @@
 #include "../core/Guard.hpp"
 #include "../core/Path.hpp"
 #include "../core/Random.hpp"
-#include "../drawing/Drawing.h"
+#include "../drawing/Palette.h"
 #include "../drawing/PaletteIndex.h"
 #include "../entity/Duck.h"
 #include "../entity/Staff.h"
@@ -363,7 +363,7 @@ static void ScenarioUpdateDayNightCycle()
     // Only update palette if day / night cycle has changed
     if (gDayNightCycle != currentDayNightCycle)
     {
-        UpdatePalette(gGamePalette, Drawing::PaletteIndex::pi10, 236);
+        Drawing::UpdatePalette(Drawing::gGamePalette, Drawing::PaletteIndex::pi10, 236);
     }
 }
