+++ src/openrct2/scenes/editor/EditorScene.cpp
@@ -20,7 +20,7 @@
 #include "../../actions/park/LandSetRightsAction.h"
 #include "../../audio/Audio.h"
 #include "../../core/Path.hpp"
-#include "../../drawing/Drawing.h"
+#include "../../drawing/Palette.h"
 #include "../../entity/EntityList.h"
 #include "../../entity/Guest.h"
 #include "../../entity/PatrolArea.h"
@@ -58,7 +58,7 @@ void EditorScene::Load()
     OpenEditorWindows();
     resetMainViewport();
 
-    LoadPalette();
+    Drawing::LoadPalette();
     gScreenAge = 0;
     gameState.scenarioOptions.name = LanguageGetString(STR_MY_NEW_SCENARIO);
 
@@ -312,7 +312,7 @@ void EditorScene::FinaliseMainView()
     windowManager->BroadcastIntent(Intent(INTENT_ACTION_CLEAR_TILE_INSPECTOR_CLIPBOARD));
 
     gWindowUpdateTicks = 0;
-    LoadPalette();
+    Drawing::LoadPalette();
 }
 
 void EditorScene::resetMainViewport()
