+++ src/openrct2/scenes/intro/IntroScene.cpp
@@ -17,6 +17,7 @@
 #include "../../audio/AudioMixer.h"
 #include "../../drawing/Drawing.h"
 #include "../../drawing/NewDrawing.h"
+#include "../../drawing/Palette.h"
 #include "../../drawing/PaletteIndex.h"
 #include "../../drawing/Rectangle.h"
 #include "../../scenes/SceneManager.h"
