+++ src/openrct2/scenes/title/TitleScene.cpp
@@ -19,6 +19,7 @@
 #include "../../config/Config.h"
 #include "../../core/Console.hpp"
 #include "../../drawing/Drawing.h"
+#include "../../drawing/Palette.h"
 #include "../../interface/Screenshot.h"
 #include "../../interface/Window.h"
 #include "../../network/NetworkBase.h"
@@ -153,7 +154,7 @@ void TitleScene::Tick()
         {
             gameStateUpdateLogic();
         }
-        UpdatePaletteEffects();
+        Drawing::UpdatePaletteEffects();
         // update_weather_animation();
     }
