+++ test/tests/S6ImportExportTests.cpp
@@ -22,6 +22,7 @@
 #include <openrct2/core/MemoryStream.h>
 #include <openrct2/core/String.hpp>
 #include <openrct2/drawing/Drawing.h>
+#include <openrct2/drawing/Palette.h>
 #include <openrct2/entity/EntityRegistry.h>
 #include <openrct2/entity/EntityTweener.h>
 #include <openrct2/object/ObjectManager.h>
@@ -60,7 +61,7 @@ static void GameInit(bool retainSpatialIndices)
         gameState.entities.resetEntitySpatialIndices();
 
     ResetAllSpriteQuadrantPlacements();
-    LoadPalette();
+    Drawing::LoadPalette();
     EntityTweener::get().reset();
     MapAnimations::MarkAllTiles();
     FixInvalidVehicleSpriteSizes();
